chore: move env dumper utils
